Fix cargo features
Unfortunately, there does not appear to be a good way to conditionally
enable a feature of an optional dependency only if two both of two
features are defined (in this case datetime and postgres). We were
accidentally depending on postgres anytime datetime was enabled.

To remediate, remove "postgres/with-chrono-0.4" from datetime and
enable it always when postgres is enabled, as it doesn't seem worth
introducing a separate feature for, although Cargo's limitations here
are unfortunate.