Make FromSql::from_sql_ref mandatory
Coupled with the previous change to return SqlValRef from
ConnectionMethods::query, this opens the door to greater efficiency
by avoiding an extra copy for FieldTypes which
can't be directly moved in from the SqlVal and for which the SqlVal
therefore represents an unnecessry copy.

This will make forthcoming custom type work more efficient.