Add a `butane clean` command (#46)
Clean current migration state. Deletes the current migration working
state which is generated on each build. This can be used as a
workaround to remove stale tables from the schema, as Butane does not
currently auto-detect model removals. The next build will recreate
with only tables for the extant models.

This is a workaround for #44