Add #[derive(FieldType)] (#68)
Add the ability to write `#[derive(FieldType)]` on a struct. The struct will be serialized as JSON. In the future, we may expand the capabilities here (for example, to allow deriving FieldType on an enum)

Based on #49 from @jayvdb. The majority of the credit for this feature is to him.


Co-authored-by: John Vandenberg <jayvdb@gmail.com>