Add more flexibility in type path matching in proc macro
Unfortunately because we don't have the type system available at
proc-macro time, when evaluating model fields we look for specific
paths like `Option`, `ForeignKey` etc. This change makes it easy to
check for the base name as well as fully qualified name and allows
fully-qualified names for `Option`, `Many`, and `ForeignKey`. It still
does not support aliased names.

Fixes #115