Add Many.load_ordered (#131)
Loads the values of the `Many` sorted. Accomplished through new `query`/`load_query` methods which `load` is also rewritten in terms of. 

These other new methods are not public at this time as they have some limitations that may not be obvious from the methods signature, but we may make them public in the future if the need for other similar use cases arises.