Create dir before lock (#213)
Related to #56

Logically, this seems to be correct.
However, I dont want to assume this fixes it. I did a lot of repeats of running all the tests, and havent seen this problem re-occur. But I cant reliably reproduce it, so cant be certain this has fixed it.
The question remains, how did the migrations get into this state?