[cli] return, not exit, from get_migrations
Previously, the cli-internal method `get_migrations` would print an
error and call `std::process::exit` if no migrations directory was
found, despite the fact that it has a `Result` return type.

This caused unintended consequences in #100, which unconditionally
calls `get_migrations` on process start. It is expected that no
migrations directory would exist prior to invoking `butane init`, but
this invocation causes `butane init` to fail.

It's both less-surprising and more idiomatic rust for `get_migrations`
to return an error rather than failing.

Fixes #331, however this issue also highlights that we have a large
test gap with butane-cli. This change does not close that gap.