Support is_in check in filter!/query! (#351)
query::BoolExpr::In has existed for some time. In SQL backends (all of them currently) this translates into an IN query. We did not expose this in the convenience macros however.

This change exposes it with a synthetic is_in function supported by filter. Use like
query!(Post, title.is_in("Foo", "Bar"))

Syntax like title in ("Foo", "Bar") would arguably be more natural. We could explore that as a future enhancement, but it would be more work as currently the filter! macro parses its input using syn as syntactically valid Rust.

Addresses #345